requirements.txt
asgiref==3.2.3
Django==3.0
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
djangorestframework==3.10.3
markdown==3.1.1
drf-yasg==1.17.0
